+++ src/main/java/com/github/theredbrain/overhauleddamage/mixin/entity/damage/DamageSourcesMixin.java
@@ -1,7 +1,10 @@
 package com.github.theredbrain.overhauleddamage.mixin.entity.damage;
 
+import com.github.theredbrain.overhauleddamage.entity.UsesCustomDamageType;
 import com.github.theredbrain.overhauleddamage.registry.DamageTypesRegistry;
 import com.github.theredbrain.overhauleddamage.registry.Tags;
+import com.llamalad7.mixinextras.injector.wrapmethod.WrapMethod;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
 import net.minecraft.resources.ResourceKey;
 import net.minecraft.world.damagesource.DamageSource;
 import net.minecraft.world.damagesource.DamageSources;
@@ -11,27 +14,28 @@ import net.minecraft.world.entity.LivingEntity;
 import org.jetbrains.annotations.Nullable;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
-import org.spongepowered.asm.mixin.injection.At;
-import org.spongepowered.asm.mixin.injection.Inject;
-import org.spongepowered.asm.mixin.injection.callback.CallbackInfoReturnable;
 
 @Mixin(DamageSources.class)
 public abstract class DamageSourcesMixin {
 
 	@Shadow
-	public abstract DamageSource source(ResourceKey<DamageType> key, @Nullable Entity attacker);
+	public abstract DamageSource source(ResourceKey<DamageType> key, @Nullable Entity cause);
 
-	@Inject(method = "mobAttack", at = @At("HEAD"), cancellable = true)
-	public void overhauleddamage$mobAttack(LivingEntity attacker, CallbackInfoReturnable<DamageSource> cir) {
-		if (attacker.getType().is(Tags.ATTACKS_WITH_BASHING)) {
-			cir.setReturnValue(this.source(DamageTypesRegistry.MOB_BASHING_DAMAGE_TYPE, attacker));
-			cir.cancel();
-		} else if (attacker.getType().is(Tags.ATTACKS_WITH_PIERCING)) {
-			cir.setReturnValue(this.source(DamageTypesRegistry.MOB_PIERCING_DAMAGE_TYPE, attacker));
-			cir.cancel();
-		} else if (attacker.getType().is(Tags.ATTACKS_WITH_SLASHING)) {
-			cir.setReturnValue(this.source(DamageTypesRegistry.MOB_SLASHING_DAMAGE_TYPE, attacker));
-			cir.cancel();
+	@WrapMethod(method = "mobAttack")
+	public DamageSource overhauleddamage$wrap_mobAttack(LivingEntity mob, Operation<DamageSource> original) {
+		if (mob instanceof UsesCustomDamageType usesCustomDamageType) {
+			ResourceKey<DamageType> customDamageType = usesCustomDamageType.overhauleddamage$getCustomDamageType();
+			if (customDamageType != null) {
+				return this.source(customDamageType, mob);
+			}
 		}
+		if (mob.is(Tags.ATTACKS_WITH_BASHING)) {
+			return this.source(DamageTypesRegistry.MOB_BASHING_DAMAGE_TYPE, mob);
+		} else if (mob.is(Tags.ATTACKS_WITH_PIERCING)) {
+			return this.source(DamageTypesRegistry.MOB_PIERCING_DAMAGE_TYPE, mob);
+		} else if (mob.is(Tags.ATTACKS_WITH_SLASHING)) {
+			return this.source(DamageTypesRegistry.MOB_SLASHING_DAMAGE_TYPE, mob);
+		}
+		return original.call(mob);
 	}
 }